Display byte keys with a given key formatter
For LoggingContext to print record keys in bytes as human-readable
String, this commit introduces a new interface `KeyFormatter` that
translates keys from bytes into `String`.
This interface has a canonical implementation that reads the byte key as
a UTF-8 byte sequence.

For other places, this commit just prevents the byte keys from being
printed, or prints it as String if it's inside a test case.